serial-version/python-controller/requirements.txt
PySide6==6.6.1
pyserial==3.5
